fix tidy, circular dep
